+++ episodio.cpp
@@ -5,7 +5,7 @@
 #include <vector>
 using namespace std;
 
-Episodio::Episodio(std::string id, std::string nombre, int temporada, int numEpisodio, std::vector<std::string> generos, int calificacion, int duracion, std::string fechaEstreno)
+Episodio::Episodio(std::string id, std::string nombre, int temporada, int numEpisodio, std::vector<std::string> generos, int calificacion, double duracion, std::string fechaEstreno)
 : Video(id, nombre, generos, calificacion, duracion, fechaEstreno) {
     this->temporada = temporada;
     this->numEpisodio = numEpisodio;
